Random questions & pre.determined number of questions
This is actually supposed to be a part of a project I'm currently calling Algolab.

The user gets a few questions regarding JavaScript. If the answer is correct, there will be a green border. If false, the border will be red.
At the end of the quiz, the user will see how many questions were answered correctly.